# Conflicts:
#	backend/src/main/java/top/minecode/service/staff/StaffSubTaskCheckService.java

+++ backend/src/main/java/top/minecode/domain/task/TaskType.java
@@ -45,7 +45,7 @@ public enum TaskType {
 
     public static double getPrice(TaskType taskType) {
         double value = Double.parseDouble(taskType.toString().substring(2));
-        return value / 100;
+        return (int) value / 100;
     }
 
     public static List<TaskType> getAll() {
